Move Water and N uptake arbitration into separate class
